Create additional modules for sales analyst to improve organization and SRP
